requirements.txt
# Web UI
streamlit>=1.32.0

# AI / Agentic framework
langchain>=0.2.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.2.0

# Document parsing
pandas>=2.2.0
openpyxl>=3.1.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
